fixed some methods in interfaces

+++ app/src/main/java/com/blstream/neverendingstory/Interfaces/ITask.java
@@ -8,9 +8,9 @@ public interface ITask {
 
     String getName();
 
-    int getInitialTime();
+    long getInitialTime();
 
-    int getElapsedTime();
+    long getElapsedTime();
 
     ISingleTaskService getService();
